refactor(vm): virtual machine controller to api endpoints
refactor vm controller to api endpoints

Closes #809
