+++ src/Sepes.RestApi/ApiEndpoints/Sandboxes/Create.cs
@@ -3,10 +3,6 @@ using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using Sepes.Common.Dto.Sandbox;
 using Sepes.Infrastructure.Service.Interface;
-using Sepes.RestApi.ApiEndpoints.Base;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net.Mime;
 using System.Threading;
 using System.Threading.Tasks;
